RDM-2736: Move to external API
Add explicit external API package
Move v2 CaseController to external API
